New command: \artRGE
Argument: 'true' or 'false'.

If true, a box is added on the bottom left of the frontpage with a statement about exam copy. It needs to be signed by the student if s/he wants to take a photo of the exam during the correction visit.

Usage: add the command after the \begin{document}, right before \namebox{}.
**Do not leave an empty line between the two commands.

Requirements: Needs the file 'art88.tex'.